meltria: move DatasetRecord class into meltria package
