+++ meltria/loader.py
@@ -8,6 +8,29 @@ import pandas as pd
 from tsdr import tsdr
 
 
+class DatasetRecord:
+    """A record of dataset"""
+    chaos_comp: str     # chaos-injected component
+    chaos_type: str     # injected chaos type
+    metrics_file: str   # path of metrics file
+    data_df: pd.DataFrame
+
+    def __init__(self, chaos_type: str, chaos_comp: str, metrics_file: str, data_df: pd.DataFrame):
+        self.chaos_comp = chaos_comp
+        self.chaos_type = chaos_type
+        self.metrics_file = metrics_file
+        self.data_df = data_df
+
+    def chaos_case(self) -> str:
+        return f"{self.chaos_comp}/{self.chaos_type}"
+
+    def chaos_case_file(self) -> str:
+        return f"{self.metrics_file} of {self.chaos_case()}"
+
+    def metrics_names(self) -> list[str]:
+        return list(self.data_df.columns)
+
+
 def load_dataset(metrics_files: list[str], exclude_middleware_metrics: bool = False) -> pd.DataFrame:
     """ Load metrics dataset
     """
